backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
pydantic==2.10.2
pydantic[email]==2.10.2
python-dotenv==1.0.1
alembic==1.14.0
email-validator==2.2.0
cryptography==43.0.3
requests==2.32.3
redis==5.2.0
celery==5.4.0
python-decouple==3.8
pydantic-settings==2.6.1
aiosmtplib==3.0.1
jinja2==3.1.4
